requirements.txt
openpyxl==2.6.3
selenium==3.6.0
cryptography==2.1.4
pandas==1.0.4
beautifulsoup4==4.9.1
PyYAML==5.3.1
